backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0
neo4j>=5.20.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
spacy>=3.7.0
langdetect>=1.0.9
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.27.0
pdfplumber>=0.11.0
python-docx>=1.1.0
reportlab>=4.2.0
websockets>=12.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4
supabase>=2.4.0
langchain-community>=0.2.0
ragas>=0.1.0
wandb>=0.17.0
